It's fun to commit lots of code :> This allows NQ to compile under: -Wall -Wpointer-arith -Wcast-qual -Wsign-compare -Werror -Winline, QW will come when I feel like going through this pain again...

+++ src/nq/client.h
@@ -206,7 +206,7 @@ typedef struct {
 
 	char        levelname[40];			// for display on solo scoreboard
 	int         viewentity;				// cl_entitites[cl.viewentity] = player
-	int         maxclients;
+	Uint8		maxclients;
 	int         gametype;
 
 // refresh related state
